Work on post detail screen

+++ shared/src/commonMain/kotlin/dev/bitbakery/boilerplate/post/inject/PostComponent.kt
@@ -40,6 +40,6 @@ interface PostComponent {
 
     @IntoMap
     @Provides
-    fun providePostDetailViewModel(factory: (postId: Int) -> PostDetailViewModel): Pair<KClass<out ViewModel>, Any> =
+    fun providePostDetailViewModel(factory: (postId: Long) -> PostDetailViewModel): Pair<KClass<out ViewModel>, Any> =
         PostDetailViewModel::class to PostDetailViewModel.Factory(factory)
 }
